Move AccountOut to AccountIn.kt and rename it to AccountClasses.kt for clarity.
